requirements.txt
greenlet==3.3.2
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48
typing_extensions==4.15.0
fastapi==0.135.1
uvicorn==0.42.0
pydantic==2.12.5
python-dotenv==1.0.0